==> Sales_Recon/src/dplyr/2-3_Join_Material.R <==
library(dplyr)
library(readr)
library(stringr)
library(here)


# Path
path <- here("Sales_Recon")


# Functions
main <- function() {
  # Filenames
  input_0 <- here(path, "output", "2-2. Join Tax invoice and SAP billing.csv")
  input_1 <- here(path, "output", "1-1. Tax invoice all.csv")
  input_2 <- here(path, "output", "1-2. SAP billing summary.csv")
  output_file <- here(path, "output", "2-3. 입출고 비교.csv")

  # Read data
  df <- read_csv(input_0, col_types = cols(.default = col_character()))
  df_1 <- read_csv(input_1, col_types = cols(.default = col_character()))
  df_2 <- read_csv(input_2, col_types = cols(.default = col_character()))

  # Process data
  # Select columns and Remove duplicate rows ----
  key_columns <- c("고객명", "sold_to_party", "customer_pn_rev")
  df_1 <- df_1 |>
    select(c(all_of(key_columns), "공장명", "ship_to_party")) |>
    tidyr::unite("temp", key_columns, sep = "_", remove = FALSE)

  df_2 <- df_2 |>
    filter(!is.na(.data$고객명)) |>
    select(!c("qty", "amt", "avg_billing_price")) |>
    tidyr::unite("temp", key_columns, sep = "_", remove = FALSE)

  # Remove duplicate rows ----
  df_1a <- df_1 |>
    filter(!duplicated(pick("temp"))) |>
    select(!c("temp"))

  df_2a <- df_2 |>
    filter(!duplicated(pick("temp"))) |>
    select(!c("temp"))

  df_2b <- df_2 |>
    filter(duplicated(pick("temp"))) |>
    select(!c("temp"))

  # Join two dataframes ----
  df <- df |>
    left_join(df_2a, by = key_columns) |>
    left_join(df_1a, by = key_columns)

  # bind_rows ----
  df <- df |> bind_rows(df_2b)

  # Arrange rows by column values ----
  df <- df |>
    arrange(pick(
      "고객명",
      "sold_to_party",
      "customer_pn_rev",
      "plant",
      "profit_center",
      "material_number"
    ))

  # Write data
  write_excel_csv(df, output_file, na = "0")
  print("A file is created")
}

main()
